refactored//cleaned api
